파일이동
routers/shop ->
routes/shop 으로

community에 임시로 작성해둔 /shop 요청 라우터들 이동 완료
