+++ routes/community.js
@@ -590,10 +590,6 @@ router.post('/exchange/insert', upload.single('img'), async (req, res) => {
 });
 
 
-
-// ---------------------------shop
-
-
 // 쇼핑몰아이템 삽입
 router.post('/insertShopItem', upload.single('img'), async (req, res) => {
   // const userId = req.user._id;
@@ -657,205 +653,7 @@ router.post('/daily/editShopItem/:itemId', upload.single('img'), async (req, res
   }
 });
 
-router.get('/shop', async (req, res) => {
-  let posts;
-  if (req.query.nextId) {
-    posts = await db.collection('shop').find({ _id: { $gt: new ObjectId(req.query.nextId) } }).limit(8).toArray();
-  } else {
-    posts = await db.collection('shop').find({}).limit(8).toArray();
-  }
-  // res.render('write.ejs', { posts })
-  res.json({
-    flag: true,
-    message: '성공적으로 상품을 가져왔습니다.',
-    posts
-  });
-});
 
-// 상품정보 불러오기(초기 8개, 더보기 시 8개 추가)
-// 상품 태그별로 보여주기 feed
-router.get('/shop/feed', async (req, res) => {
-  let posts;
-  if (req.query.nextId) {
-    posts = await db.collection('shop').find({ _id: { $gt: new ObjectId(req.query.nextId) }, tag: 'feed' }).limit(8).toArray();
-  } else {
-    posts = await db.collection('shop').find({ tag: 'feed' }).limit(8).toArray();
-  }
-  console.log(posts);
-  res.render('write.ejs', { posts })
-  // res.json({
-  //   flag: true,
-  //   message: '성공적으로 상품을 가져왔습니다.(feed)',
-  //   posts
-  // });
-});
-
-// 장바구니 추가
-router.post('/plusCart', async (req, res) => {
-  const title = req.body.title;
-  const price = req.body.price;
-  const postId = req.body.postId;
-  const count = req.body.count;
-  try {
-    const user = req.user._id;
-    await db.collection('cart').insertOne({ title, price, count, postId: new ObjectId(postId), user });
-    res.json({
-      flag: true,
-      message: '장바구니 추가 완료'
-    });
-  } catch (err) {
-    console.error(err);
-  }
-});
-
-// 수량 1개씩 추가 버튼
-router.post('/plusCount', async (req, res) => {
-  const postId = req.body.postId;
-  const user = req.user._id;
-  try {
-    await db.collection('cart').updateOne({ postId, user }, { $inc: { count: 1 }});
-    const result = await db.collection('cart').findOne({ postId, user });
-    const count = result.count;
-    res.json({
-      flag: true,
-      message: '카운트 +1 성공',
-      count
-    });
-  } catch (err) {
-    console.error(err);
-  }
-});
-router.post('/minusCount', async (req, res) => {
-  const postId = req.body.postId;
-  const user = req.user._id;
-  try {
-    const data = await db.collection('cart').findOne({ postId, user });
-    if (data.count === '1' ) {
-      throw new Error('1 이하로 감소시킬 수 없습니다');
-    } else {
-      await db.collection('cart').updateOne({ postId, user }, { $inc: { count: -1 }});
-      const result = await db.collection('cart').findOne({ postId, user });
-      const count = result.count;
-      res.json({
-        flag: true,
-        message: '카운트 -1 성공',
-        count
-      });
-    }
-  } catch (err) {
-    console.error(err);
-    res.json({
-      flag: false,
-      message: err.message
-    });
-  }
-});
-
-
-// 상품 상세페이지 가져오기
-router.get('/shop/:postId', async (req, res) => {
-  try {
-    const postId = req.params.postId;
-    const itemDetail = await db.collection('shop').findOne({ _id: new ObjectId(postId) });
-    // 상세정보 더미 만들고 내려줘야함 (지금없음)
-    // const itemDetail = await db.collection('shop').findOne({ _id: new ObjectId(postId) });
-    res.json({
-      flag: true,
-      message: '상세정보 불러오기 성공',
-      itemDetail,
-  
-    });
-  } catch (err) {
-    console.error(err);
-  }
-});
-
-// 상품 상세페이지_리뷰 가져오기
-router.get('/shop/review/:postId', async (req, res) => {
-  try {
-    const postId = req.params.postId;
-    const itemReview = await db.collection('review').find({ postId: new ObjectId(postId) }).toArray();
-    res.json({
-      flag: true,
-      message: '리뷰 불러오기 성공',
-      itemReview
-    });
-  } catch (err) {
-    console.error(err);
-  }
-});
-
-// 상품 상세페이지_리뷰 작성하기
-router.post('/shop/reviewInsert/:postId', upload.single('img'), async (req, res) => {
-  try {
-    const postId = req.params.postId;
-    console.log(postId);
-    const brand = req.body.brand;
-    const title = req.body.title;
-    const content = req.body.content;
-    const date = req.body.date; 
-    const imgUrl = req.file?.location || '';
-    const imgKey = req.file?.key || '';
-    await db.collection('review').insertOne({ brand, title, content, date, postId, imgUrl, imgKey });
-    res.json({
-      flag: true,
-      message: '리뷰 등록 완료'
-    });
-  } catch (err) {
-    console.error(err);
-    res.json({
-      flag: false,
-      message: '리뷰 등록 실패'
-    });
-  };
-});
-
-// 상품 상세페이지_Q&A 가져오기
-router.get('/shop/qna/:postId', async (req, res) => {
-  try {
-    const postId = req.params.postId
-    const itemQna = await db.collection('qna').find({ postId: new ObjectId(postId) }).toArray();
-    res.json({
-      flag: true,
-      message: 'Q&N 불러오기 성공',
-      itemQna
-    });
-  } catch (err) {
-    console.error(err);
-  }
-});
-
-// 상품 Q&A 작성하기
-router.post('/shop/qna/:postId', async (req, res) => {
-  try {
-    const postId = req.params.postId;
-    const title = req.body.title;
-    const content = req.body.content;
-    await db.collection('qna').insertOne({ title, content, postId: new ObjectId(postId), status: '답변대기' });
-    res.json({
-      flag: true,
-      message: 'Q&N 등록 완료'
-    });
-  } catch (err) {
-    console.error(err);
-  }
-});
-
-// 상품 Q&A 답변하기
-router.patch('/shop/qnaComment/:qnaPostId', async (req, res) => {
-  try {
-    const qnaPostId = req.params.qnaPostId;
-    const answer = req.body.answer;
-    await db.collection('qna').updateOne({ _id: new ObjectId(qnaPostId) }, { $set: { answer } });
-    await db.collection('qna').updateOne({ _id: new ObjectId(qnaPostId) }, { $set: { status: '답변완료' } });
-    res.json({
-      flag: true,
-      message: 'Q&N 답변 등록 완료'
-    });
-  } catch (err) {
-    console.error(err);
-  }
-});
